requirements.txt
numpy
biopython
freesasa
pytorch
pandas
tqdm
beautifulsoup4